// front/src/app/api-client/model/meteo-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeteoJourDTO } from './meteo-jour-dto';


/**
 * Météo d\'aujourd\'hui et de demain
 */
export interface MeteoDTO { 
    lieu?: string;
    /**
     * Aujourd\'hui puis demain
     */
    jours?: Array<MeteoJourDTO>;
}